requirements.txt
streamlit
youtube_transcript_api
langchain_text_splitters
langchain_openai 
azure-identity
langchain_core
dotenv
langchain-community 
faiss-cpu